backend/requirements.txt
aiohttp-cors==0.7.0
aiohttp==3.5.4
importlib-metadata==0.19 ; python_version < '3.8'
numpy==1.17.0
pandas==0.25.1
python-dateutil==2.8.0
pyyaml==5.1.2
